fix: 🐛 Disallow scaling to zero px dimensions (#6)
In a downstream filter this was causing the error: DOMException: Failed to execute 'drawImage' on 'CanvasRenderingContext2D': The image argument is a canvas element with a width or height of 0.
This was caused by a user trying to make a slide layer so small it was invisible. The node renderer handled this fine, but the canvas renderer crashed.